+++ src/Bridge/SurrealDb/CHANGELOG.md
@@ -6,6 +6,9 @@ CHANGELOG
 
  * Normalize the `endpoint` URL and tolerate a trailing slash
  * [BC BREAK] Rename the `$endpointUrl` constructor/factory argument to `$endpoint`
+ * Introduce a `StoreFactory`
+ * [BC BREAK] Add support for `ScopingHttpClient` in `Store`
+ * [BC BREAK] The `endpointUrl` parameter for `Store` has been removed
 
 0.1
 ---